services/notification-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.27.2
dapr-ext-fastapi==0.13.0
python-dotenv==1.0.1